+++ src/cli/dev/dev-server/function-manager.ts
@@ -75,6 +75,12 @@ export class FunctionManager {
     try {
       return await promise;
     } finally {
+      if (!this.starting.has(name) && this.running.has(name)) {
+        // We can end up here if user called `stopAll()` while one of the processes is being initialized.
+        // In this case we need to kill the process, to avoid zombies running around.
+        this.running.get(name)?.process.kill();
+        this.running.delete(name);
+      }
       this.starting.delete(name);
     }
   }
@@ -98,9 +104,14 @@ export class FunctionManager {
     return this.waitForReady(name, runningFunc);
   }
 
+  reload(functions: BackendFunction[]): void {
+    this.stopAll();
+    this.functions = new Map(functions.map((f) => [f.name, f]));
+  }
+
   stopAll(): void {
     for (const [name, { process }] of this.running) {
-      this.logger.log(`[dev-server] Stopping function: ${name}`);
+      this.logger.log(`Stopping function: ${name}`);
       process.kill();
     }
     this.running.clear();
@@ -113,9 +124,7 @@ export class FunctionManager {
   }
 
   private spawnFunction(func: BackendFunction, port: number): ChildProcess {
-    this.logger.log(
-      `[dev-server] Spawning function "${func.name}" on port ${port}`,
-    );
+    this.logger.log(`Spawning function "${func.name}" on port ${port}`);
 
     const process = spawn("deno", ["run", "--allow-all", WRAPPER_PATH], {
       env: {
@@ -148,14 +157,17 @@ export class FunctionManager {
     });
 
     process.on("exit", (code) => {
-      this.logger.log(
-        `[dev-server] Function "${name}" exited with code ${code}`,
-      );
+      // `code === null` happens when process is terminated by a signal.
+      // In this case I'm assuming that it's happening as part of the reload mechanism.
+      // In other words there is no need to log `code` information, if we ourselve killed the process.
+      if (code !== null) {
+        this.logger.log(`Function "${name}" exited with code ${code}`);
+      }
       this.running.delete(name);
     });
 
     process.on("error", (error) => {
-      this.logger.error(`[dev-server] Function "${name}" error:`, error);
+      this.logger.error(`Function "${name}" error:`, error);
       this.running.delete(name);
     });
   }